feat(combat): add skill cast time with charging bar and combat fixes
- Add cast time to all skills (0.5-1.0s) with ATB hold during charge
- Broadcast casting_skill and cast_progress to client via proto
- Client CombatATBPanel shows gold charging bar with skill name
- Block player movement during casting (IsCasting check)
- Migrate monster config from JSON to Luban xlsx tables
- Sync player HP changes from combat back to authoritative state
- Fix monster names in combat logs (use real names)
- Add InCombat flag to prevent repeated collision detection
- Add OnDisengage callback for monster AI resume after combat
- Add diagnostic collision logging

+++ clinetcsharp/Scripts/NetworkManager.cs
@@ -343,7 +343,7 @@ namespace ClinetCSharp
                     case MessageId.GatewayHeartbeatRsp:
                     {
                         var rsp = Gateway.HeartbeatResponse.Parser.ParseFrom(data);
-                        ServerTime = rsp.ServerTime;
+                        ServerTime = (uint)rsp.ServerTime;
                         break;
                     }
 
@@ -388,7 +388,7 @@ namespace ClinetCSharp
                         {
                             GatewayToken = rsp.GatewayToken;
                             MaxRoleCount = rsp.MaxRoleCount;
-                            ServerTime = rsp.ServerTime;
+                            ServerTime = (uint)rsp.ServerTime;
                             Roles = new List<Login.RoleBrief>(rsp.Roles);
                         }
                         SelectServerResponse?.Invoke(rsp);
@@ -431,8 +431,8 @@ namespace ClinetCSharp
                         if (rsp.Code == Common.ErrorCode.Success)
                         {
                             CurrentMapName = rsp.MapName;
-                            SpawnGridX = rsp.SpawnX;
-                            SpawnGridY = rsp.SpawnY;
+                            SpawnGridX = (int)rsp.SpawnX;
+                            SpawnGridY = (int)rsp.SpawnY;
                         }
                         ChangeMapResponse?.Invoke(rsp);
                         break;
